requirements.txt
altair==5.5.0
attrs==25.3.0
blinker==1.9.0
cachetools==6.1.0
certifi==2025.7.14
charset-normalizer==3.4.2
click==8.2.1
cloudpickle==3.1.1
contourpy==1.3.3
cycler==0.12.1
deprecation==2.1.0
fonttools==4.59.0
gitdb==4.0.12
GitPython==3.1.45
idna==3.10
Jinja2==3.1.6
joblib==1.5.1
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
kiwisolver==1.4.8
lightgbm==4.6.0
llvmlite==0.44.0
MarkupSafe==3.0.2
matplotlib==3.10.3
narwhals==1.48.1
numba==0.61.2
numpy==2.2.6
nvidia-nccl-cu12==2.27.6
packaging==25.0
pandas==2.3.1
pillow==11.3.0
plotly==6.2.0
protobuf==6.31.1
pyarrow==21.0.0
pydeck==0.9.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
referencing==0.36.2
requests==2.32.4
rpds-py==0.26.0
scikit-learn==1.7.1
scipy==1.16.0
seaborn==0.13.2
setuptools==80.10.2
shap==0.48.0
six==1.17.0
slicer==0.0.8
smmap==5.0.2
streamlit==1.47.1
streamlit-modal==0.1.2
tenacity==9.1.2
threadpoolctl==3.6.0
toml==0.10.2
tornado==6.5.1
tqdm==4.67.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
watchdog==6.0.0
wheel==0.46.3
xgboost==3.0.2
